requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
APScheduler==3.10.4
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
lxml==4.9.3
